pyproject.toml
[tool.poetry]
name = "lit-review"
version = "0.1.0"
description = ""
authors = ["Tom Kerby <46465501+tjkerby@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
ipykernel = ">=6.27,<7.0"
requests = ">=1.0"
pandas = ">=2.0"
neo4j = ">=5.0"
tqdm = ">=4.56"
pymupdf4llm = ">=0.0.15,<0.0.18"
langchain = ">=0.3.14,<0.4.0"
python-dotenv = ">=1.0.1,<2.0.0"
langchain-neo4j = ">=0.2.0,<0.3.0"
torch = ">=2.5.1,<3.0.0"
sentence-transformers = ">=3.3.1,<4.0.0"
accelerate = ">=1.3.0,<2.0.0"
bitsandbytes = ">=0.45.0,<0.46.0"
ipywidgets = ">=8.1.5,<9.0.0"
jupyter = ">=1.1.1,<2.0.0"
langchain-huggingface = ">=0.1.2,<0.2.0"
einops = ">=0.8.0,<0.9.0"
langchain-experimental = ">=0.3.4,<0.4.0"
transformers = "^4.48.3"
langchain-ollama = "^0.3.2"
matplotlib = "^3.10"
langgraph = "^0.2.74"
langgraph-checkpoint-sqlite = "^2"
aiosqlite = "^0.20.0"
wikipedia = "^1.4.0"
langchain-openai = "^0.3.12"
arxiv = "^2.2.0"
outlines = "^0.2.3"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
